Explain prefix sum and the digits to array linkage problem
Sorting the digits needs to translate to sorting the array. This is done by using prefix sum to make sure no. of occurrences and thus indices are recorded. Using counting sort guarantees stability so that the digits can be linked to appropriate positions for the actual elements in the sorted array. Do a dry run to see this.